Ajout de la couche formulaire Refacto validation Utilisation d'entier pour le stockages des montants et arrondi à l affichage en float

+++ conf/evolutions/default/1.sql
@@ -6,8 +6,8 @@
 create table accounting_row (
   id                        bigint not null,
   label                     varchar(255),
-  amount                    float,
-  personal_withdrawal       float,
+  amount                    integer,
+  personal_withdrawal       integer,
   row_type                  varchar(7),
   category_id               bigint,
   treasury_id               bigint,
